AP11-33: add confidence score and gff file option
The confidence score from the glimmer files is now captured in the
ProteinLocation class.

The main class accepts a -p optional parameter to output to gff file.

+++ src/au/org/intersect/protein_generator/domain/ProteinLocation.java
@@ -1,5 +1,7 @@
 package au.org.intersect.protein_generator.domain;
 
+import java.math.BigDecimal;
+
 public class ProteinLocation
 {
     public static final String FORWARD = "+";
@@ -9,7 +11,7 @@ public class ProteinLocation
     private int startIndex;
     private int length;
     private String direction;
-    private Integer confidenceScore;
+    private BigDecimal confidenceScore;
 
     public ProteinLocation(String name, int startIndex, int length, String direction)
     {
@@ -20,7 +22,7 @@ public class ProteinLocation
         this.confidenceScore = null;
     }
 
-    public ProteinLocation(String name, int startIndex, int length, String direction, Integer confidenceScore)
+    public ProteinLocation(String name, int startIndex, int length, String direction, BigDecimal confidenceScore)
     {
         this.name = name;
         this.startIndex = startIndex;
